feat: update total work rounds to use f64 for improved precision in stats calculations

+++ src-tauri/src/commands.rs
@@ -481,7 +481,7 @@ pub fn stats_get_heatmap(db: State<'_, DbState>) -> Result<HeatmapStats, String>
     })?;
     Ok(HeatmapStats {
         entries,
-        total_rounds: raw.completed_work_sessions as u32,
+        total_rounds: raw.completed_work_sessions,
         total_focus_secs: raw.total_work_secs as u64,
         longest_streak: streak.longest,
     })
@@ -900,7 +900,7 @@ pub struct DetailedStats {
 #[derive(serde::Serialize)]
 pub struct HeatmapStats {
     pub entries: Vec<queries::HeatmapEntry>,
-    pub total_rounds: u32,
+    pub total_rounds: f64,
     pub total_focus_secs: u64,
     pub longest_streak: u32,
 }
